Add security hardening guidance (#50)
Also move the security section to a separate doc,
because it is becoming too large to be in the snap description.

I can't think of any specific guidance for the snap,
since it does not provide any configuration
and already comes strictly confined with a minimal set of interfaces.
It feels like a good idea to explicitly note this in the docs however,
and we can also link to official Tailscale hardening docs
so the user has a good starting point.